Adjust variable length output array sizes
Update D3D12 repolay code generator to apply the same variable sized
output array adjustment that is used for Vulkan. For API calls that have
an optional output array parameter and a porinter to an integer size
parameter, when the array parameter is NULL, store the retrieved size in
the calling objects info struct, and then when the API call is made with
a non-NULL array parameter, specify the size stored in the info struct
when allocating the output array.